11_util/totalcmd python/03_GroundWater Ussage/06_KakaoMap/kako_local/pyproject.toml
[project]
name = "mcp-kakao-local"
version = "0.1.0"
description = "MCP for Kakao Local"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx[http2]>=0.28.1",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
mcp_kakao_local = "mcp_kakao_local:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]

[tool.ruff]
line-length = 100
indent-width = 2
